pyproject.toml
[project]
name = "invoiceninja-mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "awiseguy",email = "arvin@wise-it.info"}
]
requires-python = ">=3.10"
dependencies = [
    "fastmcp (>=2.12.4,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "ruff (>=0.13.3,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
